Rettet fra int til long efter afstemning fra holdet

+++ src/main/java/org/example/adventurexp/repositories/IEquipmentRepository.java
@@ -4,8 +4,8 @@ import org.example.adventurexp.model.Equipment;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
-public interface IEquipmentRepository extends JpaRepository<Equipment, Integer>{
+public interface IEquipmentRepository extends JpaRepository<Equipment, Long>{
 
     // Returnerer en liste af Equipment ud fra ActivityID
-    List<Equipment> findByActivityId(int activityId);
+    List<Equipment> findByActivityId(long activityId);
 }
